silence queue enumeration "normal" errors
Connecting to Cups may fail because it is not running yet. As Puppet
tries to enumerate instances before applying the catalog, no amount of
resource ordering will allow to start the service before we try to
connect to it.

Furthermore, the desired state might be to stop the service or to not
manage it at all.

Therefore, purge_unmanaged_queues => true may fail for reasons we can't
fix, and scare the users with a big red error message (or worse, lead to
ignore-the-errors mindset).

Thus, silence the error while enumerating Cups queues.