avoid ipptool errors when Cups is slow to start
Problem statement: on some OSes (most notably RHEL7 and derivatives), the
Cups service may have started and not listen on the IPP port yet. This causes
a problem when this module (re)starts the service and expects to be able to
add/remove printers with ipptool right after.

Thus, modify the standard cups.socket service on RHEL7 through a drop-in
to also manage the IPP port. This way, systemd will accept the incoming
connections and hand them off to Cups once it is really started. This is
mostly a kludge for a 3rd party bug, but still required to make the
module behave reliably.

Fix some cases of #35.

Breaking changes:
- Bump the Puppet minimum version to 4.10.0 for Hiera 5.
- Depend on puppetlabs/stdlib and camptocamp/systemd (and its dependency
  puppetlabs/inifile).